adding controllers and refactoring

+++ src/main/java/com/cart/carrinhoDeCompras/repositories/ItemsRepository.java
@@ -3,8 +3,6 @@ package com.cart.carrinhoDeCompras.repositories;
 import com.cart.carrinhoDeCompras.entities.Items;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface ItemsRepository extends JpaRepository<Items, Long> {
+public interface ItemsRepository extends JpaRepository<Items, Integer> {
     public Items findByItemName(String itemName);
 }
